Alteracoes no banco de dados nas tabelas de FuncionarioServico, OrdemVenda, OrdemVendaProduto e OrdemVendaServico para adequacao as regras de negocio Correcao de erros nas pages de cadastrovenda (Falta de converter, ondelete e commands) Implementada a funcao de cadastro de venda no dataservice Implementado a funcao de cadastro de venda na viewmodel cadastrovenda Ajustes na viewmodel de cadastrovenda para preencher os campos necessarios na hora de salvar

+++ Carro/Carro/Carro/Models/OrdemVenda.cs
@@ -27,7 +27,7 @@ namespace Carro.Models
 
         public int ParcelasPagas { get; set; }
 
-        public float Valor { get; set; }
+        public decimal Valor { get; set; }
 
         public DateTime Registro { get; set; }
